+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/TxBody.hs
@@ -108,7 +108,7 @@ import Cardano.Ledger.Shelley.TxBody (ShelleyEraTxBody (..), Wdrl (Wdrl), unWdrl
 import Cardano.Ledger.ShelleyMA.Timelocks (ValidityInterval (..))
 import Cardano.Ledger.ShelleyMA.TxBody (ShelleyMAEraTxBody (..))
 import Cardano.Ledger.TxIn (TxIn (..))
-import Cardano.Ledger.Val (Val (..), decodeMint, encodeMint)
+import Cardano.Ledger.Val (Val (..))
 import Control.DeepSeq (NFData (..))
 import Data.Sequence.Strict (StrictSeq)
 import qualified Data.Sequence.Strict as StrictSeq
@@ -465,7 +465,7 @@ encodeTxBodyRaw
       !> encodeKeyedStrictMaybe 6 atbrUpdate
       !> encodeKeyedStrictMaybe 8 bot
       !> Omit null (Key 14 (To atbrReqSignerHashes))
-      !> Omit (== mempty) (Key 9 (E encodeMint atbrMint))
+      !> Omit (== mempty) (Key 9 (To atbrMint))
       !> encodeKeyedStrictMaybe 11 atbrScriptIntegrityHash
       !> encodeKeyedStrictMaybe 7 atbrAuxDataHash
       !> encodeKeyedStrictMaybe 15 atbrTxNetworkId
@@ -503,7 +503,7 @@ instance
         ofield
           (\x tx -> tx {atbrValidityInterval = (atbrValidityInterval tx) {invalidBefore = x}})
           From
-      bodyFields 9 = field (\x tx -> tx {atbrMint = x}) (D decodeMint)
+      bodyFields 9 = field (\x tx -> tx {atbrMint = x}) From
       bodyFields 11 = ofield (\x tx -> tx {atbrScriptIntegrityHash = x}) From
       bodyFields 14 = field (\x tx -> tx {atbrReqSignerHashes = x}) From
       bodyFields 15 = ofield (\x tx -> tx {atbrTxNetworkId = x}) From